+++ lib/commanded/process_managers/process_manager.ex
@@ -6,6 +6,10 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   It handles events and dispatches commands in response. Process managers have
   state that can be used to track which aggregates are being orchestrated.
 
+  Process managers can be used to implement long-running transactions by
+  following the saga pattern. This is a sequence of commands and their
+  compensating commands which can be used to rollback on failure.
+
   Use the `Commanded.ProcessManagers.ProcessManager` macro in your process
   manager module and implement the callback functions defined in the behaviour:
 
@@ -13,6 +17,9 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   - `c:handle/2`
   - `c:apply/2`
   - `c:error/3`
+
+  Please read the [Process managers](process-managers.html) guide for more
+  detail.
 
   ### Example
 
@@ -146,6 +153,8 @@ defmodule Commanded.ProcessManagers.ProcessManager do
 
   ### Example
 
+  Define a process manager with `:strong` consistency:
+
       defmodule ExampleProcessManager do
         use Commanded.ProcessManagers.ProcessManager,
           application: ExampleApp,
@@ -153,8 +162,33 @@ defmodule Commanded.ProcessManagers.ProcessManager do
           consistency: :strong
       end
 
-  Please read the [Process managers](process-managers.html) guide for more
-  detail.
+  ## Dynamic application
+
+  A process manager's application can be provided as an option to `start_link/1`.
+  This can be used to start the same process manager multiple times, each using a
+  separate Commanded application and event store.
+
+  ### Example
+
+  Start an process manager for each tenant in a multi-tenanted app, guaranteeing
+  that the data and processing remains isolated between tenants.
+
+      for tenant <- [:tenant1, :tenant2, :tenant3] do
+        {:ok, _app} = MyApp.Application.start_link(name: tenant)
+        {:ok, _handler} = ExampleProcessManager.start_link(application: tenant)
+      end
+
+  Typically you would start the event handlers using a supervisor:
+
+      children =
+        for tenant <- [:tenant1, :tenant2, :tenant3] do
+          {ExampleProcessManager, application: tenant}
+        end
+
+      Supervisor.start_link(children, strategy: :one_for_one)
+
+  The above example requires three named Commanded applications to have already
+  been started.
   """
 
   alias Commanded.ProcessManagers.FailureContext
@@ -237,10 +271,10 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   @doc """
   Called when a command dispatch or event handling returns an error.
 
-  The `c:error/3` function allows you to control how command dispatch and event
-  handling failures are handled. The function is passed the error (e.g.
-  `{:error, :failure}`), the failed command (during failed dispatch) or failed
-  event (during failed event handling), and a failure context struct (see
+  The `c:error/3` function allows you to control how event  handling and command
+  dispatch and failures are handled. The function is passed the error (e.g.
+  `{:error, :failure}`), the failed event (during failed event handling) or
+  failed command (during failed dispatch), and a failure context struct (see
   `Commanded.ProcessManagers.FailureContext` for details).
 
   The failure context contains a context map you can use to pass transient state
@@ -257,6 +291,16 @@ defmodule Commanded.ProcessManagers.ProcessManager do
     the requested delay (in milliseconds). Context is a map as described in
     `{:retry, context}` above.
 
+  - `{:stop, reason}` - stop the process manager with the given reason.
+
+  For event handling failures, when failure source is an event, you can also
+  return:
+
+  - `:skip` - to skip the problematic event. No commands will be dispatched.
+
+  For command dispatch failures, when failure source is a command, you can also
+  return:
+
   - `{:skip, :discard_pending}` - discard the failed command and any pending
     commands.
 
@@ -264,11 +308,9 @@ defmodule Commanded.ProcessManagers.ProcessManager do
     dispatching any pending commands.
 
   - `{:continue, commands, context}` - continue dispatching the given commands.
-    This allows you to retry the failed command, modify it and retry, drop it,
+    This allows you to retry the failed command, modify it and retry, drop it
     or drop all pending commands by passing an empty list `[]`. Context is a map
     as described in `{:retry, context}` above.
-
-  - `{:stop, reason}` - stop the process manager with the given reason.
 
   """
   @callback error(
@@ -276,11 +318,12 @@ defmodule Commanded.ProcessManagers.ProcessManager do
               failure_source :: command | domain_event,
               failure_context :: FailureContext.t()
             ) ::
-              {:retry, context :: map()}
+              {:continue, commands :: list(command), context :: map()}
+              | {:retry, context :: map()}
               | {:retry, delay :: non_neg_integer(), context :: map()}
+              | :skip
               | {:skip, :discard_pending}
               | {:skip, :continue_pending}
-              | {:continue, commands :: list(command), context :: map()}
               | {:stop, reason :: term()}
 
   alias Commanded.Event.Handler
@@ -301,10 +344,11 @@ defmodule Commanded.ProcessManagers.ProcessManager do
       @name name
 
       def start_link(opts \\ []) do
+        application = Keyword.get(opts, :application, @application)
         module_opts = Keyword.drop(@opts, [:application, :name])
         opts = Handler.start_opts(__MODULE__, module_opts, opts, [:event_timeout, :idle_timeout])
 
-        ProcessRouter.start_link(@application, @name, __MODULE__, opts)
+        ProcessRouter.start_link(application, @name, __MODULE__, opts)
       end
 
       @doc """
@@ -319,8 +363,10 @@ defmodule Commanded.ProcessManagers.ProcessManager do
 
       """
       def child_spec(opts) do
+        application = Keyword.get(opts, :application, @application)
+
         default = %{
-          id: {__MODULE__, @name},
+          id: {__MODULE__, application, @name},
           start: {__MODULE__, :start_link, [opts]},
           restart: :permanent,
           type: :worker
@@ -351,6 +397,33 @@ defmodule Commanded.ProcessManagers.ProcessManager do
       def error({:error, reason}, _command, _failure_context), do: {:stop, reason}
     end
   end
+
+  @doc """
+  Get the identity of the current process instance.
+
+  This must only be called within a process manager's `handle/2` or `apply/2`
+  callback function.
+
+  ## Example
+
+      defmodule ExampleProcessManager do
+        use Commanded.ProcessManagers.ProcessManager,
+          application: MyApp.Application,
+          name: __MODULE__
+
+        def interested?(%ProcessStarted{uuids: uuids}), do: {:start, uuids}
+
+        def handle(%IdentityProcessManager{}, %ProcessStarted{} = event) do
+          # Identify which uuid is associated with the current instance from the
+          # list of uuids in the event.
+          uuid = Commanded.ProcessManagers.ProcessManager.identity()
+
+          # ...
+        end
+      end
+
+  """
+  defdelegate identity, to: Commanded.ProcessManagers.ProcessManagerInstance
 
   def compile_config(module, opts) do
     application = Keyword.get(opts, :application)